requirements/base.txt
Django==1.10.4
django-auth-abakus==1.1.0
django-crispy-forms==1.6.1
psycopg2==2.6.2
django-environ==0.4.1
raven==5.32.0
djangorestframework==3.5.3
djangorestframework-camel-case==0.2.0
